@@ -1,7 +1,7 @@
 import {Component} from "@angular/core";
 @Component({
   selector: 'app-root',
-  template:`
+  template: `
       <h1>{{name}}</h1>
       
       <button (click)="changeName()">Change name</button>
